Add files via upload
This analysis is to discover approaches to identify the potential buyers and focus more on them to opt for their courses from the given data. The data consists of customer visit their website, the time they spend and the details they file and add they watch etc. The below are the approaches and steps performed to achieve the final solution.